Refactor Toggle
