+++ packages/react/src/toggle/Toggle.tsx
@@ -1,12 +1,13 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { NOOP } from '../utils/noop';
-import { useComponentRenderer } from '../utils/useComponentRenderer';
+import { useControlled } from '../utils/useControlled';
+import { useEventCallback } from '../utils/useEventCallback';
+import { useRenderElement } from '../utils/useRenderElement';
 import type { BaseUIComponentProps } from '../utils/types';
 import { CompositeItem } from '../composite/item/CompositeItem';
 import { useToggleGroupContext } from '../toggle-group/ToggleGroupContext';
-import { useToggle } from './useToggle';
+import { useButton } from '../use-button/useButton';
 
 /**
  * A two-state button that can be on or off.
@@ -20,7 +21,7 @@ const Toggle = React.forwardRef(function Toggle(
 ) {
   const {
     className,
-    defaultPressed = false,
+    defaultPressed: defaultPressedProp = false,
     disabled: disabledProp = false,
     form, // never participates in form validation
     onPressedChange: onPressedChangeProp,
@@ -31,18 +32,31 @@ const Toggle = React.forwardRef(function Toggle(
     ...otherProps
   } = props;
 
+  const value = valueProp ?? '';
+
   const groupContext = useToggleGroupContext();
 
   const groupValue = groupContext?.value ?? [];
 
-  const { disabled, pressed, getRootProps } = useToggle({
+  const defaultPressed = groupContext ? undefined : defaultPressedProp;
+
+  const disabled = (disabledProp || groupContext?.disabled) ?? false;
+
+  const [pressed, setPressedState] = useControlled({
+    controlled: groupContext && value ? groupValue?.indexOf(value) > -1 : pressedProp,
+    default: defaultPressed,
+    name: 'Toggle',
+    state: 'pressed',
+  });
+
+  const onPressedChange = useEventCallback((nextPressed: boolean, event: Event) => {
+    groupContext?.setGroupValue?.(value, nextPressed, event);
+    onPressedChangeProp?.(nextPressed, event);
+  });
+
+  const { getButtonProps, buttonRef } = useButton({
+    disabled,
     buttonRef: forwardedRef,
-    defaultPressed: groupContext ? undefined : defaultPressed,
-    disabled: (disabledProp || groupContext?.disabled) ?? false,
-    onPressedChange: onPressedChangeProp ?? NOOP,
-    pressed: groupContext && valueProp ? groupValue?.indexOf(valueProp) > -1 : pressedProp,
-    setGroupValue: groupContext?.setGroupValue ?? NOOP,
-    value: valueProp ?? '',
   });
 
   const state: Toggle.State = React.useMemo(
@@ -53,13 +67,22 @@ const Toggle = React.forwardRef(function Toggle(
     [disabled, pressed],
   );
 
-  const { renderElement } = useComponentRenderer({
-    propGetter: getRootProps,
-    render: render ?? 'button',
-    ref: forwardedRef,
+  const renderElement = useRenderElement('button', props, {
     state,
-    className,
-    extraProps: otherProps,
+    ref: buttonRef,
+    props: [
+      {
+        'aria-pressed': pressed,
+        onClick(event: React.MouseEvent) {
+          const nextPressed = !pressed;
+          setPressedState(nextPressed);
+          onPressedChange(nextPressed, event.nativeEvent);
+        },
+        ref: buttonRef,
+      },
+      otherProps,
+      getButtonProps,
+    ],
   });
 
   return groupContext ? <CompositeItem render={renderElement()} /> : renderElement();
@@ -69,29 +92,43 @@ export { Toggle };
 
 namespace Toggle {
   export interface State {
+    /**
+     * Whether the toggle is currently pressed.
+     */
     pressed: boolean;
     /**
-     * Whether the component should ignore user interaction.
+     * Whether the toggle should ignore user interaction.
      */
     disabled: boolean;
   }
 
-  export interface Props
-    extends Partial<
-        Pick<
-          useToggle.Parameters,
-          'pressed' | 'defaultPressed' | 'disabled' | 'onPressedChange' | 'value'
-        >
-      >,
-      Omit<BaseUIComponentProps<'button', State>, 'value'> {
+  export interface Props extends Omit<BaseUIComponentProps<'button', State>, 'value'> {
     /**
-     * The label for the Toggle.
+     * Whether the toggle button is currently active.
      */
-    'aria-label'?: React.AriaAttributes['aria-label'];
+    pressed?: boolean;
     /**
-     * An id or space-separated list of ids of elements that label the Toggle.
+     * The default pressed state. Use when the component is not controlled.
+     * @default false
      */
-    'aria-labelledby'?: React.AriaAttributes['aria-labelledby'];
+    defaultPressed?: boolean;
+    /**
+     * Whether the component should ignore user interaction.
+     * @default false
+     */
+    disabled?: boolean;
+    /**
+     * Callback fired when the pressed state is changed.
+     *
+     * @param {boolean} pressed The new pressed state.
+     * @param {Event} event The corresponding event that initiated the change.
+     */
+    onPressedChange?: (pressed: boolean, event: Event) => void;
+    /**
+     * A unique string that identifies the component when used
+     * inside a ToggleGroup.
+     */
+    value?: string;
   }
 }
 
@@ -100,14 +137,6 @@ Toggle.propTypes /* remove-proptypes */ = {
   // │ These PropTypes are generated from the TypeScript type definitions. │
   // │ To update them, edit the TypeScript types and run `pnpm proptypes`. │
   // └─────────────────────────────────────────────────────────────────────┘
-  /**
-   * The label for the Toggle.
-   */
-  'aria-label': PropTypes.string,
-  /**
-   * An id or space-separated list of ids of elements that label the Toggle.
-   */
-  'aria-labelledby': PropTypes.string,
   /**
    * @ignore
    */
