+++ packages/react/src/utils/useRenderElement.tsx
@@ -113,7 +113,7 @@ export namespace useRenderElement {
     /**
      * The render prop or React element to override the default element.
      */
-    render:
+    render?:
       | undefined
       | ComponentRenderFn<React.HTMLAttributes<any>, State>
       | React.ReactElement<Record<string, unknown>>;
